+++ R/webforest-package.R
@@ -7,5 +7,6 @@
 #' @importFrom jsonlite toJSON
 #' @importFrom rlang %||% abort warn inform is_scalar_character is_scalar_double
 #' @importFrom cli cli_abort cli_warn cli_inform
+#' @importFrom checkmate assert_flag assert_number assert_string assert_choice assert_subset assert_character
 ## usethis namespace: end
 NULL